chore: extract mailchimp signup into a separate function
